feat(various): add transactions route in the api

+++ src/app/func.ts
@@ -1,7 +1,8 @@
 import {loginSucceed, loginFailed, loginRequest, logout} from "../features/auth/auth-slice";
-import {Dispatch} from "@reduxjs/toolkit";
 import {getUserDataFailed, getUserDataSucceed} from "../features/user/user-slice";
-export const logUserIn = ({ email, password }: {email: string, password: string}, dispatch: Dispatch) => {
+import {EditUserFunc, GetUserDataFunc, LogUserInFunc, LogUserOutFunc} from "./func-types";
+
+export const logUserIn: LogUserInFunc = ({ email, password }, dispatch, keepLoggedIn) => {
     dispatch(loginRequest());
 
     const uri = `${import.meta.env.VITE_API_URL}/user/login`;
@@ -27,8 +28,12 @@ export const logUserIn = ({ email, password }: {email: string, password: string}
         .then(result => {
             const { token } = result.body;
             if (token) {
-                // set token in local storage
-                localStorage.setItem("token", token);
+                if (keepLoggedIn) {
+                    localStorage.setItem('token', token);
+                } else {
+                    sessionStorage.setItem('token', token);
+                }
+
                 dispatch(loginSucceed({token}));
             } else {
                 dispatch(loginFailed(result.message));
@@ -38,12 +43,12 @@ export const logUserIn = ({ email, password }: {email: string, password: string}
             dispatch(loginFailed(error));
         })
 }
-
-export const logUserOut = (dispatch: Dispatch) => {
+export const logUserOut: LogUserOutFunc = (dispatch) => {
     localStorage.removeItem("token");
+    sessionStorage.removeItem("token");
     dispatch(logout());
 }
-export const getUserData = (token: string, dispatch: Dispatch) => {
+export const getUserData: GetUserDataFunc = (token, dispatch) => {
     const uri = `${import.meta.env.VITE_API_URL}/user/profile`;
 
     const myHeaders = new Headers({
@@ -60,9 +65,7 @@ export const getUserData = (token: string, dispatch: Dispatch) => {
     fetch(uri, requestOptions)
         .then(response => response.json())
         .then(result => {
-            console.log(result);
             if (result.body) {
-                console.log(result.body);
                 const { id, email, firstName, lastName } = result.body;
 
                 dispatch(getUserDataSucceed({
@@ -71,6 +74,49 @@ export const getUserData = (token: string, dispatch: Dispatch) => {
                     firstName,
                     lastName
                 }));
+            } else {
+                dispatch(getUserDataFailed(result.message));
+            }
+        })
+        .catch(error => {
+            dispatch(getUserDataFailed(error));
+        })
+}
+export const editUserData: EditUserFunc = (data, dispatch, resolveCallback) => {
+    const uri = `${import.meta.env.VITE_API_URL}/user/profile`;
+    const token = localStorage.getItem("token") ?? sessionStorage.getItem("token") ?? null;
+
+    const myHeaders = new Headers({
+        "Content-Type": "application/json",
+        "Authorization": `Bearer ${token}`
+    });
+
+    const raw = JSON.stringify({
+        firstName: data.firstName,
+        lastName: data.lastName
+    });
+
+    const requestOptions: RequestInit = {
+        method: "PUT",
+        headers: myHeaders,
+        body: raw,
+        redirect: "follow"
+    }
+
+    fetch(uri, requestOptions)
+        .then(response => response.json())
+        .then(result => {
+            if (result.body) {
+                const { id, email, firstName, lastName } = result.body;
+
+                dispatch(getUserDataSucceed({
+                    id,
+                    email,
+                    firstName,
+                    lastName
+                }));
+
+                resolveCallback();
             } else {
                 dispatch(getUserDataFailed(result.message));
             }
